Move API types to their own folder
