+++ src/io/react-native/react-native-io.js
@@ -4,7 +4,7 @@ import { isReactNative } from 'detect-bundler'
 import { makeReactNativeFolder } from 'disklet'
 import { base64 } from 'rfc4648'
 
-import { type EdgeIo } from '../../index.js'
+import { type EdgeIo } from '../../types/types.js'
 import { HmacDRBG, hashjs } from '../../util/crypto/external.js'
 import {
   Socket,
