+++ src/modules/account/account-reducer.js
@@ -9,7 +9,7 @@ import {
   type EdgeWalletInfo,
   type EdgeWalletInfoFull,
   type EdgeWalletStates
-} from '../../index.js'
+} from '../../types/types.js'
 import { ethereumKeyToAddress } from '../../util/crypto/external.js'
 import { type RootAction } from '../actions.js'
 import { hasCurrencyPlugin } from '../currency/currency-selectors.js'
