+++ src/modules/storage/repoFolder.js
@@ -5,7 +5,7 @@ import {
   type DiskletFile,
   type DiskletFolder,
   type EdgeIo
-} from '../../index.js'
+} from '../../types/types.js'
 import { decrypt, encrypt } from '../../util/crypto/crypto.js'
 import { utf8 } from '../../util/encoding.js'
