+++ src/modules/storage/storage-actions.js
@@ -1,6 +1,6 @@
 // @flow
 
-import { type EdgeWalletInfo } from '../../index.js'
+import { type EdgeWalletInfo } from '../../types/types.js'
 import { base58, base64 } from '../../util/encoding.js'
 import { type ApiInput } from '../root.js'
 import { loadRepoStatus, makeRepoPaths, syncRepo } from './repo.js'
