+++ src/modules/storage/storage-selectors.js
@@ -1,6 +1,6 @@
 // @flow
 
-import { type EdgeIo } from '../../index.js'
+import { type EdgeIo } from '../../types/types.js'
 import { hmacSha256 } from '../../util/crypto/crypto.js'
 import { base58, utf8 } from '../../util/encoding.js'
 import { type RootState } from '../root-reducer.js'
