+++ src/modules/storage/storage-servers.js
@@ -1,7 +1,7 @@
 // @flow
 
-import { NetworkError } from '../../error.js'
-import { type EdgeIo } from '../../index.js'
+import { NetworkError } from '../../types/error.js'
+import { type EdgeIo } from '../../types/types.js'
 
 const syncServers = [
   'https://git3.airbitz.co',
