+++ src/modules/swap/swap-api.js
@@ -3,13 +3,13 @@
 import { gt, lt } from 'biggystring'
 import { bridgifyObject } from 'yaob'
 
+import { errorNames } from '../../types/error.js'
 import {
   type EdgeSwapCurrencies,
   type EdgeSwapPluginQuote,
   type EdgeSwapQuote,
-  type EdgeSwapQuoteOptions,
-  errorNames
-} from '../../index.js'
+  type EdgeSwapQuoteOptions
+} from '../../types/types.js'
 import { fuzzyTimeout } from '../../util/promise.js'
 import { swapPluginEnabled } from '../account/account-selectors.js'
 import { type ApiInput } from '../root.js'
